redo size computation

+++ core/file_processor.py
@@ -496,10 +496,10 @@ def download_csv_from_gcs(gcs_file_path: str) -> str:
         # Get the remote file size before downloading
         blob.reload()  # Ensure we have the latest metadata
         remote_size_bytes = blob.size
-        remote_size_gb = remote_size_bytes / (1024 * 1024 * 1024)
+        remote_size_gb = float(float(remote_size_bytes) / float((1024 * 1024 * 1024)))
 
         # If the size of the file is greater than half of what is allocated to DuckDB, return exception
-        if float(remote_size_gb) > float(constants.DUCKDB_MEMORY_LIMIT.replace('GB', '')) * 0.5:
+        if remote_size_gb > float(constants.DUCKDB_MEMORY_LIMIT.replace('GB', '')) * 0.5:
             raise Exception(f"CSV file {gcs_file_path} has invalid quoting, but cannot be fixed due to size constraints. Allocate at least {remote_size_gb * 2}GB of memory to the Cloud Run function")
         
         # Download the file
